02-workflow-orchestration/dagster_project/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "dagster-zoomcamp"
version = "0.1.0"
description = "Data Engineering Zoomcamp - Dagster Orchestration"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "dagster==1.6.0",
    "dagster-webserver==1.6.0",
    "dagster-postgres==0.22.0",
    "dagster-gcp==0.22.0",
    "watchdog<4",
    "pandas>=1.0.0",
    "psycopg2-binary>=2.8.0",
    "requests>=2.25.0",
    "sqlalchemy>=1.4.0",
    "python-dotenv>=0.19.0",
    "google-cloud-storage>=2.0.0",
    "google-cloud-bigquery>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.12.0",
]

[tool.setuptools.packages.find]
where = ["."]
